requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
beautifulsoup4==4.11.1
biopython==1.80
biotite==0.35.0
brotlipy==0.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
contextlib2==21.6.0
contourpy==1.2.1
crc32c==2.4
cycler==0.12.1
deepspeed==0.5.9
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-tree==0.1.8
einops==0.8.0
fair-esm==2.0.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
formulaic==1.0.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.4.0
gemmi==0.5.8
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.8.0
hjson==3.1.0
huggingface-hub==0.12.0
idna==2.10
importlib_metadata==7.1.0
interface-meta==1.3.0
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.5
libclang==18.1.1
lifelines==0.28.0
lightgbm==3.3.5
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.6.3
ml_collections==0.1.1
mock==5.1.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.0
ninja==1.11.1.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
obonet==0.3.1
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.3
pickle5==0.0.11
Pillow==9.3.0
protobuf==3.19.6
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.24.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.5
subword-nmt==0.3.8
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tfrecord==1.14.1
threadpoolctl==3.5.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.4.0.post0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.12.0
urllib3==1.25.11
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xgboost==1.7.3
yarl==1.9.4
zipp==3.12.0
